+++ tests/test_misc.py
@@ -8,6 +8,7 @@ from homematicip.EventHook import EventHook
 from homematicip.base.helpers import bytes2str, detect_encoding
 from homematicip.base.enums import DeviceType
 
+
 def event_hook_handler2(mustBe2):
     assert mustBe2 == 2
 
@@ -51,8 +52,9 @@ def test_bytes2str():
     with pytest.raises(TypeError):
         assert bytes2str(44) == testString
 
+
 def test_auto_name_enum():
-    
+
     assert DeviceType.from_str("PUSH_BUTTON") == DeviceType.PUSH_BUTTON
     assert DeviceType.from_str(None) == None
     assert DeviceType.from_str("I_DONT_EXIST", DeviceType.DEVICE) == DeviceType.DEVICE